# improve performance with ties