@palday 2023-08-15
@andreasnoack playing around a bit I discovered that things are dramatically faster if you presort instead of repeatedly calling `partialsort!`. Doing this, I can leave the original linear search in place (for now) and still see a dramatic performance increase. I'll update the benchmarks later, but it passes the sinusoid example you proposed and gives a result much closer to 0.5.4 in the plot.