@andreasnoack 2023-08-21
When you timed, did you compare to current master or https://github.com/JuliaStats/Loess.jl/pull/76?