@palday 2023-08-21
> When you timed, did you compare to current master or #76?

The released versions -- I did `add Loess@version`