@palday 2023-08-23
> Ran the timings in [#76 (comment)](https://github.com/JuliaStats/Loess.jl/pull/76#issuecomment-1690542618). Looks like the two versions are comparable in the random case but that this PR is faster when there are ties so let's go with this one. Might be worth leaving a comment in the code with a link to this issue in case somebody in the future compares the implementation to the paper and thinks there is a potential for speedup via `partialsort!`.

done!